Cmd code review feedback
